+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 - Support of verified domains for the storage nodes (#2280)
 - `neofs-lens storage status` CLI command (#2550)
 - Human-readable output of objects' creation timestamp to `neofs-cli container list-objects` (#2653)
+- Ability to preset P2PNotary and NeoFSAlphabet roles to validators at the FS chain's genesis (#2643)
 
 ### Fixed
 - `neofs-cli netmap netinfo` documentation (#2555)